requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.8.1+cpu 
torchvision==0.9.1+cpu 
torchaudio==0.8.1 
tensorboard==2.5.0
Cython==3.0a6
matplotlib==3.1.2
deap==1.3.0
h5py==2.10.0
ipyparallel==6.3.0
llvmlite==0.35.0
mpmath==1.1.0
numpy-financial==1.0.0
numpysane==0.21
openpyxl==3.0.6
pandas==0.25.3
pip==21.0.1
Platypus-Opt==1.0.3
ply==3.11
sklearn==0.0
sympy==1.8
tabulate==0.8.7
tqdm==4.40.2
wirerope==0.4.2
yabox==1.0.4
